add react routes
